Re-recreate #837 but moving from UWP w/o submodule changes
