+++ ReactWindows/ReactNative.Net46.Tests/ReactNative.Net46.Tests.csproj
@@ -105,6 +105,8 @@
     <Reference Include="Newtonsoft.Json">
       <HintPath>..\packages\Newtonsoft.Json.9.0.1\lib\net45\Newtonsoft.Json.dll</HintPath>
     </Reference>
+    <Reference Include="System.Net.Http" />
+    <Reference Include="System.Net.Http.WebRequest" />
     <Reference Include="System.Windows.Presentation" />
     <Reference Include="System.Xaml" />
     <Reference Include="WindowsBase">
@@ -118,22 +120,10 @@
     <Compile Include="Bridge\Queue\MessageQueueThreadTests.cs" />
     <Compile Include="Internal\Constants\DictionaryHelpers.cs" />
     <Compile Include="Internal\DispatcherHelpers.cs" />
-    <Compile Include="Internal\MockInvocationHandler.cs" />
-    <Compile Include="Internal\MockPromise.cs" />
-    <Compile Include="Internal\MockReactInstance.cs" />
     <Compile Include="Internal\MockViewManager.cs" />
     <Compile Include="Modules\NetInfo\NetInfoModuleTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="UIManager\AppRegistryTests.cs" />
-    <Compile Include="UIManager\DependencyObjectExtensionsTests.cs" />
-    <Compile Include="UIManager\PropertySetterTests.cs" />
-    <Compile Include="UIManager\ReactStylesDiffMapTests.cs" />
-    <Compile Include="UIManager\RootViewHelperTests.cs" />
-    <Compile Include="UIManager\ShadowNodeRegistryTests.cs" />
     <Compile Include="UIManager\UIManagerModuleTests.cs" />
-    <Compile Include="UIManager\ViewAtIndexTests.cs" />
-    <Compile Include="UIManager\ViewManagerRegistryTests.cs" />
-    <Compile Include="UIManager\ViewManagersPropertyCacheTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
