+++ ReactWindows/ReactNative.Shared.Tests/Bridge/NativeModuleRegistryTests.cs
@@ -1,4 +1,4 @@
-﻿using Microsoft.VisualStudio.TestPlatform.UnitTestFramework;
+﻿using NUnit.Framework;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using ReactNative.Bridge;
@@ -8,10 +8,10 @@ using System.Linq;
 
 namespace ReactNative.Tests.Bridge
 {
-    [TestClass]
+    [TestFixture]
     public class NativeModuleRegistryTests
     {
-        [TestMethod]
+        [Test]
         public void NativeModuleRegistry_ArgumentChecks()
         {
             var builder = new NativeModuleRegistry.Builder();
@@ -20,7 +20,7 @@ namespace ReactNative.Tests.Bridge
                 ex => Assert.AreEqual("module", ex.ParamName));
         }
 
-        [TestMethod]
+        [Test]
         public void NativeModuleRegistry_Override_Disallowed()
         {
             var builder = new NativeModuleRegistry.Builder();
@@ -28,7 +28,7 @@ namespace ReactNative.Tests.Bridge
             AssertEx.Throws<InvalidOperationException>(() => builder.Add(new OverrideDisallowedModule()));
         }
 
-        [TestMethod]
+        [Test]
         public void NativeModuleRegistry_Override_Allowed()
         {
             var registry = new NativeModuleRegistry.Builder()
@@ -39,7 +39,7 @@ namespace ReactNative.Tests.Bridge
             Assert.AreEqual(1, registry.Modules.Count());
         }
 
-        [TestMethod]
+        [Test]
         public void NativeModuleRegistry_ModuleWithNullName_Throws()
         {
             var builder = new NativeModuleRegistry.Builder();
@@ -48,7 +48,7 @@ namespace ReactNative.Tests.Bridge
                 ex => Assert.AreEqual("module", ex.ParamName));
         }
 
-        [TestMethod]
+        [Test]
         public void NativeModuleRegistry_WriteModuleDefinitions()
         {
             var registry = new NativeModuleRegistry.Builder()
