+++ ReactWindows/ReactNative.Shared.Tests/Internal/JavaScriptModuleRegistryExtensions.cs
