+++ ReactWindows/ReactNative.Shared.Tests/Internal/MockCallback.cs
