+++ ReactWindows/ReactNative.Shared.Tests/Internal/MockEvent.cs
