+++ ReactWindows/ReactNative.Shared.Tests/Internal/MockInvocationHandler.cs
