+++ ReactWindows/ReactNative.Shared.Tests/Internal/MockPromise.cs
