+++ ReactWindows/ReactNative.Shared.Tests/Internal/MockReactInstance.cs
