+++ ReactWindows/ReactNative.Shared.Tests/Modules/Core/RCTDeviceEventEmitterTests.cs
@@ -1,15 +1,14 @@
-﻿using Microsoft.VisualStudio.TestPlatform.UnitTestFramework;
+﻿using NUnit.Framework;
 using ReactNative.Modules.Core;
 
 namespace ReactNative.Tests.Modules.Core
 {
-    [TestClass]
-    public class RCTNativeAppEventEmitterTests
+    [TestFixture]
+    public class RCTDeviceEventEmitterTests
     {
-        [TestMethod]
-        public void RCTEventEmitter_InvokeTests()
+        public void RCTDeviceEventEmitter_Invoke()
         {
-            var module = new RCTNativeAppEventEmitter();
+            var module = new RCTDeviceEventEmitter();
 
             var name = default(string);
             var args = default(object[]);
@@ -22,7 +21,7 @@ namespace ReactNative.Tests.Modules.Core
             var eventName = "foo";
             var data = new object();
             module.emit(eventName, data);
-            Assert.AreEqual(nameof(RCTNativeAppEventEmitter.emit), name);
+            Assert.AreEqual(nameof(RCTDeviceEventEmitter.emit), name);
             Assert.AreEqual(2, args.Length);
             Assert.AreSame(eventName, args[0]);
             Assert.AreSame(data, args[1]);
