+++ ReactWindows/ReactNative.Shared.Tests/UIManager/AppRegistryTests.cs
