+++ ReactWindows/ReactNative.Shared.Tests/UIManager/RootViewHelperTests.cs
