+++ ReactWindows/ReactNative.Shared.Tests/UIManager/ViewAtIndexTests.cs
