+++ ReactWindows/ReactNative.Tests/ReactNative.Tests.csproj
@@ -96,48 +96,23 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Bridge\JavaScriptModuleRegistryTests.cs" />
-    <Compile Include="Bridge\PromiseTests.cs" />
     <Compile Include="Bridge\ReactInstanceTests.cs" />
-    <Compile Include="Bridge\NativeModuleBaseTests.cs" />
     <Compile Include="Bridge\Queue\MessageQueueThreadTests.cs" />
-    <Compile Include="Internal\AssertEx.cs" />
-    <Compile Include="Bridge\NativeModuleRegistryTests.cs" />
     <Compile Include="Internal\Constants\DictionaryHelpers.cs" />
     <Compile Include="Internal\DispatcherHelpers.cs" />
-    <Compile Include="Internal\JavaScriptModuleRegistryExtensions.cs" />
-    <Compile Include="Internal\MockCallback.cs" />
-    <Compile Include="Internal\MockPromise.cs" />
-    <Compile Include="Internal\MockReactInstance.cs" />
-    <Compile Include="Internal\MockInvocationHandler.cs" />
-    <Compile Include="Internal\MockEvent.cs" />
     <Compile Include="Internal\MockViewManager.cs" />
-    <Compile Include="Modules\AppState\AppStateModuleTests.cs" />
     <Compile Include="Modules\Clipboard\ClipboardModuleTests.cs" />
     <Compile Include="Modules\Clipboard\MockClipboardInstance.cs" />
-    <Compile Include="Modules\Core\JSTimersExecutionTests.cs" />
-    <Compile Include="Modules\Core\RCTDeviceEventEmitterTests.cs" />
-    <Compile Include="Modules\Core\RCTNativeAppEventEmitterTests.cs" />
     <Compile Include="Modules\Core\TimingTests.cs" />
     <Compile Include="Modules\NetInfo\NetInfoModuleTests.cs" />
-    <Compile Include="Modules\Network\NetworkingModuleTests.cs" />
-    <Compile Include="Modules\Network\TaskCancellationManagerTests.cs" />
     <Compile Include="Modules\StatusBar\StatusBarModuleTests.cs" />
     <Compile Include="Modules\Storage\AsyncStorageModuleTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ReactInstanceManagerTests.cs" />
-    <Compile Include="UIManager\AppRegistryTests.cs" />
-    <Compile Include="UIManager\ReactStylesDiffMapTests.cs" />
     <Compile Include="UIManager\Events\EventDispatcherTests.cs" />
     <Compile Include="UIManager\Events\EventTests.cs" />
     <Compile Include="UIManager\Events\RCTEventEmitterTests.cs" />
-    <Compile Include="UIManager\DependencyObjectExtensionsTests.cs" />
-    <Compile Include="UIManager\PropertySetterTests.cs" />
-    <Compile Include="UIManager\RootViewHelperTests.cs" />
-    <Compile Include="UIManager\ShadowNodeRegistryTests.cs" />
     <Compile Include="UIManager\UIManagerModuleTests.cs" />
-    <Compile Include="UIManager\ViewAtIndexTests.cs" />
-    <Compile Include="UIManager\ViewManagersPropertyCacheTests.cs" />
-    <Compile Include="UIManager\ViewManagerRegistryTests.cs" />
     <Compile Include="UnitTestApp.xaml.cs">
       <DependentUpon>UnitTestApp.xaml</DependentUpon>
     </Compile>
